Added running animation w/ BUGS
I added the running sprite animation for the squirrel.

We need to find a way to scale these sprites, we attempted doing this by
changing model width and height but the model:play() negated any
changes.

We also need to find a way to scale the acorns too.

Adding a bug to the issues for this.